damagegoal: fix assert when checking if a group has taken damage
Since the damage is determined in scheduled context DCS could have
already cleaned up the object. The fix is to check if <obj>.getByName()
actually returns an object before using the value.

We could do this another way by checking the asset when the DCS event
fires, however, since checking all objects of an asset for every HIT and
DEAD event is too expensive we can spend a little more time with an if
check which we should have been doing in the first place.

Closes: #18