check goals: fixup processing of hit and dead events
Change asset checking to a scheduled function instead of only running
when an event is triggered. This should account for objects that "burn
down". In addition on a DEAD event remove any associated death goal in
the asset as the health of the object doesn't seem to reduce to zero in
all cases.