tickets system: add basic ticket accounting system
This adds support for DCT to track a side's number of remaining tickets.
There is also the ability to limit a campaign to a fixed period of time.
In addition the campaign designer can select from 4 preset difficulties
or specify their own difficulty. If a preset difficulty is selected
this will override any custom changes you have made.

Closes: #88