handle daylight savings time on the server
Since we use os.data the local server can influence the time based on
if the server thinks daylight savings time is in effect. Add some functions
that support the ability to print time in local and zulu time. Local time
will normalize and remove the effects of daylight savings.