spawning: support airbase and airbase defense assets
This adds basic support for an airbase asset which will eventually be a
meta object that contains references to assets it 'manages'. Currently
the collection associated with the airbase is the Null collection so
nothing is currently done.

This patch also spawns any airbase defense asset associated with the
airbase. These defense assets are just like any other Static collection.

Closes: #24