Ability to always spawn a template
This change makes it such that a template definition can define the
attribute 'spawnalways' and if set to true the template will always
cause the generation of an Asset. This will be counted against the
maximum limit for the type of asset. Once all templates marked to always
spawn are processed the remaining templates will be spawned via random
slection until the region limit for the asset type is reached.

Closes: #7