asset: use composition to represent different asset types
Instead of deriving a new Asset class for each kind of asset type
just isolate the primary thing that changes which is the association
with DCS in-game objects. This collection of DCS objects can then be
operated on and delegated to without having to rewrite portions
of the Asset class over and over.

This also includes updates to the Asset API to include;
* setting initial intel state from a template
* splitting priority into individual elements

Closes: #46
Closes: #49