theater: write then rename state file
To avoid losing the old state file before the new state file is ready
move to a write new rename model so that we are all but guaranteed we
have a new state file before attempting to delete the old one. This is
done by using os.rename which on most filesystems is an atomic operation
and usually cannot fail barring hardware failure.