mission join
Allows for players to join the same mission by just being given the package
number instead of having players provide the mission information verbally.

When a member of a mission aborts they remove only them self. All other
members of the mission are unaffected until the mission times out, at which
point all members of the mission are removed and the mission is terminated.