assetmanager: remove unneeded code
Remove force flag from checkAssets as it is not used. In addition
change the timer used to os.clock() instead of timer.getTime() as
getTime() does not actually tick during a single run.

We also remove the extraneous code from handleDead() event handler
as all of that checking had already been done before we get to
the handler.