fix state saving
With the changing to composition in the Asset class, state saving was
broken. This is caused when all units of an asset were killed but the
asset is kept in the list. In addition when all units of a group were
marked as dead the group entry was not removed. This was further
compounded by changing the storage format of template data for a flat
list.

Therefore there are three changes;
1. In StaticCollection:marshal(), test if nil instead of testing
   if the table returned is empty as this nil could be returned
2. In filterDeadObjects(), if there is a units table (meaning the
   entry is a group) test if the units table is empty after filtering
   meaning the group data should not be stored
3. In Asset:marshal(), if nil is returned from the collection, the
   asset should be deleted, exactly as before all the conversions

Closes: #73
Fixes: 2070317ff04e ("modify storage of groups in templates")
Fixes: 942e4c4d27be ("asset: use composition to represent different
asset types")