remove spawndead feature
The code supported the ability to spawn dead templates in a misguided
attempt to provide atmosphere to the mission. This feature apparently
never worked correctly. Instead these assets would be saved to be
restored later but since they had no death criteria (remember they are
dead already) they would trigger the code below.

In StaticCollection:_setup():
	...
	assert(next(self._deathgoals) ~= nil,
		"runtime error: StaticCollection must have a deathgoal: "..
		self._asset.name)
end

This assert prevented the state file from being read back in and allowing
a mission to continue. The solution is to remove this code complexity
and once an Asset is considered dead it is removed from the AssetManager.

Closes: #76
Fixes: 942e4c4d27be ("asset: use composition to represent different asset types")