Revert "theater: write then rename state file"
This reverts commit 7132f94b21394c4eaaceba2429db368ded38ec62.

DCS apparently overrides various file operations or they do not operate
as documented in the LUA 5.1 manual. An example is the fact that the
file:write(blob) operation always returns `nil` even though this according
to the documentation means there was a problem writing to the file. Of
note when an error occurs there is a second parameter returned that is
suppose to contain the reason message, this value too is always `nil`.

In addition os.rename() appears non-functional on Windows platforms and
testing shows the rename of a file does not occur.