loadouts: fix payload validation for A2A weapons
Valid was getting overwritten when it checked if AG weapons were valid.
In addition since one index in total is zero this zero index is not
considered valid for ipairs(). If looping over the list of weapon types
(aa and ag for now) worked, AA would have come first making 'valid'
false. Then when the ag is checked 'valid' is set to true. Thus erasing
the invalid aa information. So just return immediately if we notice a
weapon type is invalid.