hooks: a DCS game server hook for DCT
This hook exports various information about the server as well as
providing SimpleSlotBlocker equivalent functionality.

Also, provide mission restart capability after X seconds. Currently
this relies on the server to have a list of missions setup before
hand. If the same mission is intended to be replayed set the loop
option in the server management screen and have only one mission
in the list.

Closes: #90