hooks: work around timer issues
We must work around the `DCS.getRealTime()` call as it appears this
timer continually increases when the .miz file is not changed but the
server reloads the mission. However, when the .miz is changed to a new
file this timer does get reset to zero.

Work around the issue by storing the value of the timer on mission
load end. When calculating the local time remove this value from
the local time. This should account for the timer not getting reset
consistently.