hooks: fix reference to slot when player is in spectator
When a player switches from a slot back to spectator we miss removing
the reference to the empty slot string in the player object. This
leads to us attempting to kick the player from the slot and failing
to lookup the slot's group name.

A traceback similar to the below is seen:
2020-10-04 02:14:55.519 ERROR   [DCT-HOOKS]: call to hook failed; [string "C:\Us
ers\DCS server\SavedGames\DCS.openbeta_server\Scripts/Hooks//dct-hook.lua"]:511: attempt to index field '?' (a nil value)
stack traceback:
        [string "C:\Users\DCS server\SavedGames\DCS.openbeta_server\Scripts/Hooks//dct-hook.lua"]:552: in function 'dct_call_hook'
        [string "C:\Users\DCS server\SavedGames\DCS.openbeta_server\Scripts/Hooks//dct-hook.lua"]:589: in function <[string "C:\Users\DCS server\SavedGames\DCS.openbeta_server\Scripts/Hooks//dct-hook.lua"]:588>
        [C]: in function 'pcall'
        [string "MissionEditor/loadUserScripts.lua"]:129: in function <[string "MissionEditor/loadUserScripts.lua"]:124>
        (tail call): ?