dct-hooks: fix for nil when player disconnects
When a player disconnects from the server onPlayerChangeSlot is still
called, however `net.get_player_info(id)` returns nil for the player.
This causes the following check to fail because it incorrectly assumed
the player data would be available from get_player_info(). Add the nil
check for the player entry to prevent this null dereference.