settings: split out server settings from settings function
Split out the server specific (dct.cfg) settings into its own function
so that the reading of the dct.cfg file can be done in other contexts,
such as the hooks context. This allows for common code to be written to
parse the dct.cfg file. This forces the dct.cfg parsing code to not
reference any mission environment objects as they are no longer
guarenteed to exist in the lua environment passed.