dct-hooks: fix format expected for player kick list
The kick list format in the Theater class stores the group names as keys
in the kick table. The original code assumed the names would be stored
as values. In addition the argument passed to the kick function is the
name of the group to be kicked as the mission environment has no
idea what the player's id is. Therefore, we need to use the group name
to player id table to get the id of the player that needs to be kicked.
This was not done before and was simply incorrect.

Fixes: 7f9fcbd7b680 ("hooks: efficient slot management support")