hooks: fix for lack of env available in hook environment
The 'env' table is not available in the hooks environment. So that this
code can be used in both the hooks and mission environments we need to
abstract away the direct use of 'env'. We accomplish this by using an
intermediate function that will detect which environment it is in and
choose the correct access method.