assets: marshal only alive assets
Now that we can notify when an asset dies we can atomically synchronize
asset book keeping. This removes the need to periodically check assets
so we no longer need the asset API 'checkDead'. This was always a race
condition so it needed to be gotten rid of.

We also fix a regression that was introduced in 935e4f7c509c
("reorganize: assets") that causes assets to be marshalled even if they
a dead. This would appear as an asset stating an asset saved in the
state file has no death goals, preventing the statefile from being
loaded.

Fixes: 935e4f7c509c ("reorganize: assets")