command: fix overriding with TimedCommand
When 5d3264a3c27d ("command: introduce a timed and prioritized command")
was written it introduced a timed command to be used when the global
debug was enabled. The original if check used the wrong table and so
the TimedCommand never was used. In addition we fixup some logging
messages in this commit.

Fixes: 5d3264a3c27d ("command: introduce a timed and prioritized command")