tests: cleanup tests
The two `dct.theater = t` settings are to account for when the rest of
the code uses Theater.singleton() to get the instance of the theater.
The singleton function uses whatever is set in dct.theater otherwise
it creates a new theater instances, which was leaving to multiple
readings of the templates and the test code was not referencing the same
theater instance. This appears to only be a unit test issue and not a
problem in an actual mission.