observable: allow multiple objects to observe
It appears the way lua works a function 'foo' defined in object A will have
the same address in object B if 'foo' is not overridden. This means that
objects A and B cannot observe at the same time object C. Instead we need
to index the observer list by object and not its function there by allowing
us to have multiple observers ob the same type but different instances.