assets: airbase
Fully support an actual airbase asset. The airbase will associate
player slots that are on the airbase and control if they are
joinable based on if airbase is spawned or not. The airbase will
also spawn any subordinate templates if these templates belong
to the same side as the airbase.