hooks: fix hook not detecting DCTFLAG and running
The hooks script had a bug where it was not properly getting the "DCTFLAG"
DCS mission flag value correctly. This was caused by a syntax error
combined with requesting the incorrect return type for do_rpc.

Finally, we fix up spawning of player assets and allow airbases to
determine which player assets should be spawned.

Closes: #123