utils: fix airbase indexing
It appears calling `Airbase.getID(ab)` returns a string representation of
the numerical object ID. This means we need to wrap said call in
`tonumber()` otherwise the airbase id that is stored in the player
asset will not find the correct airbase name. This allowed things
link carriers and farps to be spawned at.