hooks: fix concatenate error when player occupies special role slot
It issue occurs when a player occupies a special role slot. This causes
the slot id to be set in the player object but the groupName field
is nil, which results in the concatenation error. The solution is
to test if the player is in a special role slot.

2021-02-02 00:32:03.259 ERROR   [DCT-HOOKS]: call to hook failed; [string "C:\Users\<user>\Saved Games\DCS.openbeta\Scripts/Hooks//dct-hook.lua"]:555: attempt to concatenate local 'grpname' (a nil value)
stack traceback:
    [string "C:\Users\<user>\Saved Games\DCS.openbeta\Scripts/Hooks//dct-hook.lua"]:602: in function 'dct_call_hook'
    [string "C:\Users\<user>\Saved Games\DCS.openbeta\Scripts/Hooks//dct-hook.lua"]:639: in function <[string "C:\Users\<user>\Saved Games\DCS.openbeta\Scripts/Hooks//dct-hook.lua"]:638>
    [C]: in function 'pcall'
    [string "MissionEditor/loadUserScripts.lua"]:129: in function <[string "MissionEditor/loadUserScripts.lua"]:124>
    (tail call): ?