hooks: loop through all slots and handle kick request
We were originally only looping through the player list to handle kick
requests. But this did not work because a player can leave the slot
before the kick happens, leaving the slot in a locked state forever.
We not loop through all slots and reset the kick state regardless
if there is a player associated with the slot.