player: respond to birth events and reset state
Since `S_EVENT_LEAVE_UNIT` does not fire for occupied planes we need
to handle resetting the player slot in a different way. We do this
by listening to `S_EVENT_BIRTH` in the occupied state. In this state
we just return a new occupied state to reset the slot.