player: add timeout for empty slots
When a player leaves a slot start a timer if the slot is assigned to a
mission. Upon the expiration of the timer remove the slot from the
assigned mission. Allow the server administrator the ability to adjust
when empty slots should expire from an assigned mission. To preserve
previous behaviour slot expiration is left off and slots will be removed
when the mission expires. It is up to the administrator to setup a
shorter slot expiration.

Closes: #152