template: set .dct as priority over .stm
Reversed order of variables passed to the function to make DCT File
contents take priority over STM file contents. DCT file is the user
created file so it makes sense that variables manually defined within it
should take priority. This allows values such as the description field
to be defined in the .dct file without having to edit the .stm file.

Closes: #166