iads: fix unbounded command queue additions
Remove the queueCommand call in IADS:disableSAM as this is a recursive
call as long as the targets are in range. Causing excessive commands
to be put on the command queue.

Also, change to tracking which SAMs need to hide and doing the hiding in
the check function instead of queuing another command to do the hiding.

Finally, in OADS:onBirth change the if check to check the category
type of the object instead of detecting if the object has a getGroup
method. This is not related to excessive command queue additions.

(cherry picked from commit e865e670e7c81b2a9545e37c5070146967f0c211)
Closes: #164
Signed-off-by: Jonathan Toppins <jtoppins@users.noreply.github.com>