player: multicrew support
To better support multicrew aircraft in the Player class ejection event
handler, only track that an ejection has occurred and do not transition to
the EmptyState until an aircraft crash event occurs or the group no longer
exists. Taking the F-14 as an example this would allow the RIO to eject and
the pilot to still fly the plane and recover the aircraft. For the pilot
death event, just ignore and rely on the plane crash event.

Closes: #169