airbase: allow designer to list slots for an airbase
Some player slots cannot be autodetected as to which airbase the slot
should belong to.  If the slot is set to 'takeoff from ground' the
mission editor does not associate an airbase with this slot. This
creates a problem as these slots are locked forever. This change allows
the designer to list which slots should be associated with a particular
airbase if the slot cannot be autodetected.

Closes: #148