groupmenu: fix crash caused under specific circumstances
missionCommands.addMenu accepts multiple user arguments to the handler
function, however, it seems that behavior is not documented, and the
extra parameters can cause memory corruption under certain circumstances
when running queued commands on units that have been removed by the slot
kicker. By reducing the user arguments to one, the functions behave
correctly and DCS doesn't crash anymore when calling them. The
theater object used to be passed in the commands, but now it is injected
directly into the playerRequest method via the singleton call (as it was
done in groupmenu.lua before).