Goal: combine DCS object goals into a single file
We have not had the need to track anything other than damage taken
by DCS objects and keeping the boilerplate around in the other files
seems needless at this time.

Also move the file to the assets folder and rename it so the file
does not get confused with GOAP goal classes.