damagegoal: refine scenery object handling
ED fixed a bug where undamaged objects were considered non-existent, so
now we can (correctly) treat non-existent objects as dead. We now remove
completed goals when an asset is spawned, so if all of its parts are
dead on spawn (for any reason), it will not be a valid target for a
mission, and will be removed from the theater. These changes are aimed
at fixing a softlock caused by scenery objects being destroyed before
the theater is fully initialized (ie. by collateral damage from the
building persistence system), where they will still be considered alive
but not be able to be damaged in the world anymore (as dead objects will
be considered non-existent and won't emit any new death events).

Closes: #224