blasteffects: track impacts instead of scenery damage
It seems on some maps attempting to keep track of which scenery got
destroyed generates a rather large amount of data. Instead of doing this
simply track the impacts of weapons and replay the impacts on restart.
This requires less tracking and we can age out the impacts after a
number of restarts.