assets: rewrite asset class hierarchy to use composition
Rewrote the asset class hierarchy to use composition inside a single
Agent class. Then layered on top a goal oriented action planner to allow
agents to react to the changing game world. This single Agent class
composed of different sensors, actions, and goals should remove the need
for the class inheritance hierarchy that existed previously.

This commit only covers the Player use case. We create player slot
Templates in a custom "__builtin" region as with the new Agent model
Player templates must exist in a region otherwise it will mess up the
template proxy key retrival in the Agent that is planned.