assets: agent: allow to reference backing Template
There currently is an issue with the saved state that if the theater is
changed it usually forces a state reset. This is due to the active
assets copying too much data from the Template and only referencing this
copy during runtime. First split this data and clearly mark what needed
to be copied to the asset. Then provide a few getter functions that will
proxy called to the backing Template class if the asset doesn't have the
key being requested. This should allow for a much easier time in
changing the backing Template set in the theater and having this
reflected in already spawned/active assets/agents.