template: stop passing region to Template.fromFile
When we read in the initial Template we really do not need the Region
information and makes testing a little bit more annoying. Instead
delay the region information until the Template is added to a Region and
require the Region to call Template:joinRegion(region) method to
finalize to Template. This makes things cleaner from a flow perspective
and only only have to fixup the group and unit names of DCT objects.