checkers: add threat attribute table to common settings
Some Agents can be vulnerable to other types of agents. Keeping a static
mapping would be far too time consuming and would not scale as new agent
kinds are added. So instead we have each Agent express which DCS attributes
the Agent is most vulnerable to. This list is then used in various ways,
from filtering contacts to culling objects when attempting to find a path.