checkers: fix multiple templates pointing to same table
A bug would occur where all templates were pointing to the same default
table. This is because non-scalar values (tables, functions) are passed
by reference in LUA. The issue is this table would later be modified by
other code which caused all templates to be updated. Did a deep copy of
the default table so each template gets its own copy.