playerslots: remove need for multiple player templates
Remove the need for multiple player templates by changing the
Agent.desc.name of the template that the player Agent is pointing to.
This allows for a little optimization in template storage.  This will
need to be revisited later for tactical units as the Agent.setup() code
references the Template during setup.