Mission: replace subordinate API with custom API
The Subordinate API is really made to cater to assets. So remove the
usage of subordinate and have the Mission support its own tree API.
This allows missions to have support missions through a tree structure,
thus a mission that is a primary mission is one that doesn't have a
parent pointer set.