tickets: allow campaign config to be changed mid campaign
Before this change the whole ticket configuration would be saved to the
state file. This made modifying the ticket configuration impossible mid
campaign which makes campaign balancing an issue. With this change only
the current ticket count and the original ticket start values are saved
to the state file. All modifiers and difficulty can be modified by
stopping the server, modifying the difficulty, and reloading the state
file.