ui: implement PlayerMenu class
Uses the PlayerMenu class to implement a dynamic player menu system.
Each top level menu is treated as an Agent fact and can be individually
accessed. This allows for each top level menu to be updated independent
of the other menus. This creates a scalable way for missions to add
context actions specific to the goal the player is currently trying
to achieve.