events: add a new event for agent requests
This even can be used later to allow agents to make requests of other
agents without needing to know the details of how the agent implements
the request.

TODO: agents will need a way to express supported requests otherwise
some sort of feedback system will be needed to acknowledge the request
is supported.