airbase: disable DCS capture mechanic
Prevent the base game from determining when an airbase is captured. Also
add airbase theater setting for disabling the silencing of all airbase
ATC towers. Finally, modified the airbase sensor to take coalition
ownership of the game object when the DCT object is spawned and reset to
neutral when the DCT object is despawned.