theater: emit an init complete event to all other scripts
DCT is being used as a base for some mission designers and then they are
building on top of DCT by layering on other scripts like MOOSE to
customize their mission. To allow sequencing of these other scripts emit
an init complete event for all event handlers in the DCS mission
environment so that proper sequencing can take place.