theater: allow external scripts to control slot enabling
Mission designers are layering other scripts on top of DCT. This means
there initialization needs to be completed before players should be
allowed to join slots. To allow for this use case, a well known mission
flag can be used. The DCT hooks script will then use this flag to
determine if hooks are globally enabled or not. Finally a theater level
configuration can be set to initially set this flag true/false. With the
default being true so DCT will enable slots as normal as soon as it is
ready.