Action: allow to fail and cause replanning
Before this commit Action objects could only continue and complete.
There was no ability for an action to fail the plan without directly
calling the agent's replan method. Directly calling replan is
cumbersome and exposes too much of the Agent's API to objects that
should not be overly concerned. This allows for a much cleaner way
for Actions to fail a plan.