Agent: add an updateLocation method
Instead of having the DCSObjectSensor update agent and unit location
data for movable agents, have move related actions do this instead.
This allows the sensor to run less frequently and we only update
locations when an agent is known to have moved. We also keep the
location updates for state saving purposes.