Agent: create templates for DCS spawned groups
Originally the Template objects were not saved Agents generated from
unit group spawned outside of DCT, ex. mission placed groups. This made
it mildly annoying to do some things. Instead just save the Template in
the TemplateDB and let normal lookup happen for non-agent data.

We want to keep the agent's description table (desc) to a minimum as
this table is seralized when saving DCT state.